// frontend-recipe-app/src/app/models/user.model.ts
export interface User {
  id?: number;
  username: string;
  email: string;
  firstName: string;
  lastName: string;
  roles: string[];
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface AuthResponse {
  token: string;
  username: string;
  email: string;
  roles: string[];
  type: string;
}